+++ domain/src/main/java/com/ceiba/domain/repository/ParkingRepository.kt
@@ -5,7 +5,7 @@ import com.ceiba.domain.aggregate.Parking
 interface ParkingRepository {
     suspend fun enterVehicle(parking: Parking): Long?
 
-    suspend fun deleteVehicle(parking: Parking): Long?
+    suspend fun deleteVehicle(parking: Parking): Int?
 
     suspend fun calculateAmountParking(parking: Parking): Parking?
 }
